+++ src/test/resources/jsonprogrammedstub.json
@@ -1,7 +1,7 @@
 {
   "endpoint": "somepath/$1/with/$2",
   "responseprovider" : {
-    "class": "castalia.ProgrammedStub",
+    "class": "castalia.plugins.ProgrammedStub",
     "member": "process1"
   }
 }
